+++ firehydrant/runbooks.go
@@ -19,6 +19,21 @@ const (
 // CreateRunbookRequest is the payload for creating a service
 // URL: POST https://api.firehydrant.io/v1/runbooks
 type CreateRunbookRequest struct {
+	Name        string       `json:"name"`
+	Type        string       `json:"type"`
+	Description string       `json:"description"`
+	Owner       *RunbookTeam `json:"owner,omitempty"`
+
+	AttachmentRule map[string]interface{} `json:"attachment_rule,omitempty"`
+
+	Severities []RunbookRelation `json:"severities"`
+
+	Steps []RunbookStep `json:"steps,omitempty"`
+}
+
+// RunbookRelation associates a runbook to a type in FireHydrant (such as a severity)
+type RunbookRelation struct {
+	ID          string        `json:"id"`
 	Name        string        `json:"name"`
 	Description string        `json:"description"`
 	Owner       *RunbookTeam  `json:"owner,omitempty"`
@@ -40,10 +55,11 @@ type RunbookStep struct {
 // UpdateRunbookRequest is the payload for updating a service
 // URL: PATCH https://api.firehydrant.io/v1/runbooks/{id}
 type UpdateRunbookRequest struct {
-	Name        string        `json:"name,omitempty"`
-	Description string        `json:"description"`
-	Owner       *RunbookTeam  `json:"owner,omitempty"`
-	Steps       []RunbookStep `json:"steps,omitempty"`
+	Name           string                 `json:"name,omitempty"`
+	Description    string                 `json:"description"`
+	Owner          *RunbookTeam           `json:"owner,omitempty"`
+	Steps          []RunbookStep          `json:"steps,omitempty"`
+	AttachmentRule map[string]interface{} `json:"attachment_rule,omitempty"`
 }
 
 // RunbookResponse is the payload for retrieving a service
@@ -54,6 +70,8 @@ type RunbookResponse struct {
 	Description string        `json:"description"`
 	Owner       *RunbookTeam  `json:"owner"`
 	Steps       []RunbookStep `json:"steps"`
+
+	AttachmentRule map[string]interface{} `json:"attachment_rule"`
 
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`